Apply Wolgang Glunz updates to 1.0.4. Not compilable until the header files are moved around.
git-svn-id: https://svn.code.sf.net/p/libemf/code/trunk@46 07b56ee9-b158-4c8b-9baa-ff89d65b121f

+++ include/wine/basetsd.h
@@ -33,8 +33,11 @@ typedef unsigned short __uint16;
 typedef int          __int32;
 typedef unsigned int __uint32;
 
-typedef long long          __int64;
-typedef unsigned long long __uint64;
+// wogl
+//typedef long long          __int64;
+//typedef unsigned long long __uint64;
+typedef long          __int64;
+typedef unsigned long __uint64;
 
 #if defined(_WIN64)
